modified version of malloc01 designed to avoid a bug in effects
git-svn-id: http://svn.cri.ensmp.fr/svn/validation/trunk@1145 67b5017d-e62d-0410-a871-f72b1c62810f

+++ CoarseGrainParallelization/malloc01.c
@@ -6,7 +6,7 @@
 
 int main(int argc, char *argv[]) {
   int i, j;
-  int * a[N];
+  float * a[N];
 
   for(i=0;i<N;i++) {
     a[i] = malloc(N*4/*sizeof(float)*/);
